Move scores page into Competitions.
