+++ app/competitions/[competitionId]/scores/score-charts-card.tsx
