+++ app/competitions/[competitionId]/scores/skeleton-card.tsx
